requirements.txt
# Database
psycopg2-binary>=2.9.7

# AI/ML
google-generativeai>=0.3.2

# Configuration
python-dotenv>=1.0.0

# CLI Interface
tabulate>=0.9.0
colorama>=0.4.6

# Utilities
click>=8.1.7
pydantic>=2.5.0

# Resume parsing (recommended)
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.9.0

# Additional dependencies for better functionality
requests>=2.31.0
urllib3>=2.0.0